zaheer :
Original commit message from CVS:
zaheer :
* gst/tcp/gsttcp.c:
* gst/tcp/gsttcpclientsrc.c:
* gst/tcp/gsttcpclientsrc.h:
* gst/tcp/gsttcpserversrc.c:
- portability fix, to compile on OSX
(fixes #143146)

* sys/osxaudio/gstosxaudioelement.c:
* sys/osxaudio/gstosxaudiosink.c:
* sys/osxaudio/gstosxaudiosrc.c:
- compilation warnings on OSX
(fixes #143153)

me :
* ext/vorbis/vorbisdec.c : sign warning fixes

* gst-libs/gst/mixer/mixertrack.c : forgoten include
to define newly used G_MAXINT32, bad owen, bad

+++ gst/tcp/gsttcp.c
@@ -24,6 +24,8 @@
 #include "config.h"
 #endif
 
+#include <sys/types.h>
+#include <sys/socket.h>
 #include <netinet/in.h>
 #include <arpa/inet.h>
 #include <netdb.h>
@@ -154,7 +156,7 @@ gst_tcp_gdp_read_header (GstElement * this, int socket)
     return NULL;
   }
   if (ret != readsize) {
-    g_warning ("Wanted %d bytes, got %d bytes", readsize, ret);
+    g_warning ("Wanted %d bytes, got %d bytes", (int) readsize, (int) ret);
   }
   g_assert (ret == readsize);
